Update NN.py
Added the load and save weights method. Simple save/load weights and bias to a file to retain the training from train_SGD. If you run train_SGD and save_weights, then load weights and test on the test data, the evaluation is the exact same as the last epoch of the training set.